// src/app/models/workordermodel.ts
export class WorkorderModel {
  Workorderdate: string = '';
  Description: string = '';
  Workorderid: string = '';
  Employeename: string = '';
  Employeeid: string = '';
  Location: string = '';
  Status: string = '';
  Tag: string = '';
  Itemcount: string = '';
  Datecreated: string = '';
  Updatedby: string = '';
  DateFrom: string = '';
  DateTo: string = '';
}